+++ package.json
@@ -16,6 +16,7 @@
     "reactstrap": "^9.2.0",
     "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
+    "redux-saga": "^1.2.3",
     "styled-components": "^6.0.8",
     "uuid": "^9.0.1",
     "web-vitals": "^2.1.4"
