+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { App } from './components/index';
 import { Provider } from 'react-redux';
-import { store } from './store/index';
+import { store } from './redux-saga/store';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
