--- src/store/types.js
@@ -1,2 +1,0 @@
-export const ADD_JOKES = 'ADD_JOKES';
-export const GET_JOKES = 'GET_JOKES';
